Remove dependencies from the solana-client
