+++ wasm/solana-cli/src/utils/mod.rs
@@ -1,3 +1,4 @@
+pub mod blockhash_query;
 pub mod checks;
 pub mod memo;
 pub mod spend_utils;
