+++ wasm/solana-client/Cargo.toml
@@ -11,16 +11,13 @@ keywords = ["solana", "wasm", "client", "rpc", "playground"]
 readme = "README.md"
 
 [features]
-pubsub = ["js-sys", "web-sys", "wasm-bindgen", "wasm-bindgen-futures"]
+pubsub = ["web-sys", "wasm-bindgen", "wasm-bindgen-futures"]
 
 [dependencies]
-async-trait = "0.1"
 base64 = "0.13"
 bincode = "1.3"
 bs58 = "0.4"
-clap = "3.1.18"
 futures = "0.3"
-js-sys = { version = "0.3.65", optional = true }
 reqwest = { version = "0.11.9", features = ["json"] }
 serde = "1"
 serde_derive = "1"
