+++ wasm/solana-client/src/utils/mod.rs
@@ -3,5 +3,3 @@ pub mod rpc_config;
 pub mod rpc_filter;
 pub mod rpc_request;
 pub mod rpc_response;
-
-mod clap;
