+++ wasm/spl-token-cli/src/lib.rs
@@ -7,4 +7,5 @@ mod config;
 mod constants;
 mod output;
 mod sort;
+mod utils;
 mod wasm;
