+++ wasm/utils/solana-extra/Cargo.toml
@@ -14,7 +14,6 @@ readme = "README.md"
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
-Inflector = "0.11"
 arrayref = "0.3"
 assert_matches = "1"
 base64 = "0.13"
@@ -22,21 +21,13 @@ bincode = "1"
 borsh = "0.10"
 bs58 = "0.4"
 bytemuck = "1"
-chrono = { version = "0.4", features = ["serde"] }
-clap = "3"
-console = "0.15"
 fluvio-wasm-timer = "0.2"
-humantime = "2"
-lazy_static = "1"
-log = "0.4.17" # only for nightly builds
 num-derive = "0.3"
 num-traits = "0.2"
 num_enum = "0.5"
-semver = "1"
 serde = "1"
 serde_derive = "1"
 serde_json = "1"
-solana-frozen-abi = "~1.18"
 solana-frozen-abi-macro = "~1.18"
 solana-sdk = "~1.18"
 thiserror = "1"
